api/requirements.txt
fastapi==0.97.0
httpx==0.24.1
pydantic==1.10.8
python-dotenv==1.0.0
uvicorn==0.22.0
huggingface-hub==0.25.0
sentence_transformers==2.2.2
pgvector==0.1.8
asyncpg==0.27.0
